Create README.md
This Read Me document describes the steps to complete the project for Getting and Cleaning Data.

For script to run, you should first install and load the plyr package.

Initially, the script reads in the data from both the train and test datasets. Features are also read in to be used later to name the variables. Subject IDs are also read in.

Both training and test sets are combined to create one dataset which contains the data, labels and subject ID. I have also included a variable (which is not needed) to identify whether that record came from the training or test dataset. The resultant dataset is called full.all.

Next, the script extracts out feature names and inserts them into full.all for later use.

The following step identifies features which are means and standard deviations to produce a subset called extracted. 

The script then uses descriptive activity names to name the activities in the data set. It does this by reading in the activity labels from activity_labels.txt and merging them with the extracted dataset to produce a dataset called merged.

Finally, the script uses the ddply function (from plyr package) to generate the mean of each measurement, by subjectID and Activity.

This tidy dataset is then output to a text file called tidy.txt